Refactor ChatView
